+++ .eslintrc.js
@@ -5,11 +5,6 @@ module.exports = {
   parserOptions: {
     ecmaVersion: 2020,
   },
-  env: {
-    es6: true,
-    node: true,
-    browser: true,
-  },
   extends: [
     // Turn on on eslint recommended rules https://github.com/eslint/eslint/blob/main/conf/eslint-recommended.js
     'eslint:recommended',
@@ -17,13 +12,6 @@ module.exports = {
     'prettier',
   ],
   overrides: [
-    {
-      files: ['*.{js,mjs}'],
-      extends: [
-        // Handle prettier rules through eslint https://github.com/prettier/eslint-plugin-prettier/blob/master/eslint-plugin-prettier.js#L65
-        'plugin:prettier/recommended',
-      ],
-    },
     {
       files: ['*.{ts,tsx}'],
       parser: '@typescript-eslint/parser',
@@ -43,6 +31,8 @@ module.exports = {
         'plugin:@typescript-eslint/recommended',
         // Handle prettier rules through eslint https://github.com/prettier/eslint-plugin-prettier/blob/master/eslint-plugin-prettier.js#L65
         'plugin:prettier/recommended',
+        // add special jest rules https://github.com/jest-community/eslint-plugin-jest#rules
+        'plugin:jest/recommended',
       ],
       rules: {
         '@typescript-eslint/explicit-function-return-type': 'off',
@@ -62,15 +52,34 @@ module.exports = {
         '@typescript-eslint/require-await': 'error',
         '@typescript-eslint/no-empty-function': 'off',
         '@typescript-eslint/ban-types': 'off', // TODO: turn on
+        /** jest */
+        'jest/valid-title': 'off', // allow functions to be used as describe titles
+        'jest/no-conditional-expect': 'off', // best practice, but TODO
+        'jest/no-alias-methods': 'off', // best practice, but TODO
+        'jest/expect-expect': 'off', // sometimes we compose assertion functions
+        'jest/no-disabled-tests': process.env.CI ? 'error' : 'off', // fail CI if tests are disabled, but do not interfere with local development
+        'jest/no-focused-tests': process.env.CI ? 'error' : 'off',
       },
       overrides: [
         {
-          files: ['*.test.*'],
+          files: ['**/__tests__/**', '**/scripts/**'],
           rules: {
             'require-await': 'off',
             '@typescript-eslint/require-await': 'off',
           },
         },
+      ],
+    },
+    {
+      files: ['*.{js,mjs}'],
+      env: {
+        // Config files and possible scripts. Allow anything, we don't really care.
+        browser: true,
+        node: true,
+      },
+      extends: [
+        // Handle prettier rules through eslint https://github.com/prettier/eslint-plugin-prettier/blob/master/eslint-plugin-prettier.js#L65
+        'plugin:prettier/recommended',
       ],
     },
   ],