+++ packages/browser/src/plugins/segmentio/__tests__/normalize.test.ts
@@ -7,6 +7,20 @@ import { SegmentEvent } from '../../../core/events'
 import { JSDOM } from 'jsdom'
 import { version } from '../../../generated/version'
 
+/**
+ * Filters out the calls made for probing cookie availability
+ */
+const ignoreProbeCookieWrites = (
+  fn: jest.SpyInstance<
+    string | undefined,
+    [
+      name: string,
+      value: string | object,
+      options?: cookie.CookieAttributes | undefined
+    ]
+  >
+) => fn.mock.calls.filter((c) => c[0] !== 'ajs_cookies_check')
+
 describe('before loading', () => {
   let jsdom: JSDOM
 
@@ -51,13 +65,24 @@ describe('before loading', () => {
       window.localStorage.clear()
     }
   })
+
   describe('#normalize', () => {
     let object: SegmentEvent
+    let defaultCtx: any
+    const withSearchParams = (search?: any) => {
+      object.context = { page: { search } }
+    }
 
     beforeEach(() => {
       cookie.remove('s:context.referrer')
+      defaultCtx = {
+        page: {
+          search: '',
+        },
+      }
       object = {
         type: 'track',
+        context: defaultCtx,
       }
     })
 
@@ -94,17 +119,15 @@ describe('before loading', () => {
     })
 
     it('should always add .anonymousId even if .userId is given', () => {
-      const object: SegmentEvent = { userId: 'baz', type: 'track' }
+      object.userId = 'baz'
       normalize(analytics, object, options, {})
       assert(object.anonymousId?.length === 36)
     })
 
     it('should accept anonymousId being set in an event', async () => {
-      const object: SegmentEvent = {
-        userId: 'baz',
-        type: 'track',
-        anonymousId: '👻',
-      }
+      object.userId = 'baz'
+      object.anonymousId = '👻'
+
       normalize(analytics, object, options, {})
       expect(object.anonymousId).toEqual('👻')
     })
@@ -115,15 +138,16 @@ describe('before loading', () => {
     })
 
     it('should not rewrite context if provided', () => {
-      const ctx = {}
+      const ctx = defaultCtx
       const obj = { ...object, context: ctx }
       normalize(analytics, obj, options, {})
       expect(obj.context).toEqual(ctx)
     })
 
-    it('should copy .options to .context', () => {
+    it('should overwrite options with context if context does not exist', () => {
       const opts = {}
       const obj = { ...object, options: opts }
+      delete obj.context
       normalize(analytics, obj, options, {})
       assert(obj.context === opts)
       assert(obj.options == null)
@@ -172,6 +196,7 @@ describe('before loading', () => {
 
     it('should not replace .locale if provided', () => {
       const ctx = {
+        ...defaultCtx,
         locale: 'foobar',
       }
       const obj = { ...object, context: ctx }
@@ -180,10 +205,9 @@ describe('before loading', () => {
     })
 
     it('should add .campaign', () => {
-      jsdom.reconfigure({
-        url: 'http://localhost?utm_source=source&utm_medium=medium&utm_term=term&utm_content=content&utm_campaign=name',
-      })
-
+      withSearchParams(
+        'utm_source=source&utm_medium=medium&utm_term=term&utm_content=content&utm_campaign=name'
+      )
       normalize(analytics, object, options, {})
 
       assert(object)
@@ -197,10 +221,7 @@ describe('before loading', () => {
     })
 
     it('should decode query params', () => {
-      jsdom.reconfigure({
-        url: 'http://localhost?utm_source=%5BFoo%5D',
-      })
-
+      withSearchParams('?utm_source=%5BFoo%5D')
       normalize(analytics, object, options, {})
 
       assert(object)
@@ -210,9 +231,7 @@ describe('before loading', () => {
     })
 
     it('should guard against undefined utm params', () => {
-      jsdom.reconfigure({
-        url: 'http://localhost?utm_source',
-      })
+      withSearchParams('?utm_source')
 
       normalize(analytics, object, options, {})
 
@@ -223,10 +242,7 @@ describe('before loading', () => {
     })
 
     it('should guard against empty utm params', () => {
-      jsdom.reconfigure({
-        url: 'http://localhost?utm_source=',
-      })
-
+      withSearchParams('?utm_source=')
       normalize(analytics, object, options, {})
 
       assert(object)
@@ -236,10 +252,7 @@ describe('before loading', () => {
     })
 
     it('only parses utm params suffixed with _', () => {
-      jsdom.reconfigure({
-        url: 'http://localhost?utm',
-      })
-
+      withSearchParams('?utm')
       normalize(analytics, object, options, {})
 
       assert(object)
@@ -248,9 +261,7 @@ describe('before loading', () => {
     })
 
     it('should guard against short utm params', () => {
-      jsdom.reconfigure({
-        url: 'http://localhost?utm_',
-      })
+      withSearchParams('?utm_')
 
       normalize(analytics, object, options, {})
 
@@ -260,13 +271,14 @@ describe('before loading', () => {
     })
 
     it('should allow override of .campaign', () => {
-      jsdom.reconfigure({
-        url: 'http://localhost?utm_source=source&utm_medium=medium&utm_term=term&utm_content=content&utm_campaign=name',
-      })
+      withSearchParams(
+        '?utm_source=source&utm_medium=medium&utm_term=term&utm_content=content&utm_campaign=name'
+      )
 
       const obj = {
         ...object,
         context: {
+          ...defaultCtx,
           campaign: {
             source: 'overrideSource',
             medium: 'overrideMedium',
@@ -287,10 +299,20 @@ describe('before loading', () => {
       assert(obj.context.campaign.name === 'overrideName')
     })
 
-    it('should add .referrer.id and .referrer.type (cookies)', () => {
-      jsdom.reconfigure({
-        url: 'http://localhost?utm_source=source&urid=medium',
+    it('should allow override of .search with object', () => {
+      withSearchParams({
+        someObject: 'foo',
       })
+
+      normalize(analytics, object, options, {})
+      assert(object)
+      assert(object.context)
+      assert(object.context.campaign === undefined)
+      assert(object.context.referrer === undefined)
+    })
+
+    it('should add .referrer.id and .referrer.type (cookies)', () => {
+      withSearchParams('?utm_source=source&urid=medium')
 
       normalize(analytics, object, options, {})
       assert(object)
@@ -307,9 +329,7 @@ describe('before loading', () => {
     })
 
     it('should add .referrer.id and .referrer.type (cookieless)', () => {
-      jsdom.reconfigure({
-        url: 'http://localhost?utm_source=source&urid=medium',
-      })
+      withSearchParams('utm_source=source&urid=medium')
       const setCookieSpy = jest.spyOn(cookie, 'set')
       analytics = new Analytics(
         { writeKey: options.apiKey },
@@ -323,15 +343,11 @@ describe('before loading', () => {
       expect(object.context.referrer.id).toEqual('medium')
       assert(object.context.referrer.type === 'millennial-media')
       expect(cookie.get('s:context.referrer')).toBeUndefined()
-      expect(setCookieSpy).not.toHaveBeenCalled()
+      expect(ignoreProbeCookieWrites(setCookieSpy).length).toBe(0)
     })
 
     it('should add .referrer.id and .referrer.type from cookie', () => {
       cookie.set('s:context.referrer', '{"id":"baz","type":"millennial-media"}')
-
-      jsdom.reconfigure({
-        url: 'http://localhost?utm_source=source',
-      })
 
       normalize(analytics, object, options, {})
 
@@ -347,10 +363,6 @@ describe('before loading', () => {
         's:context.referrer',
         '{"id":"medium","type":"millennial-media"}'
       )
-
-      jsdom.reconfigure({
-        url: 'http://localhost',
-      })
 
       normalize(analytics, object, options, {})
       assert(object)
@@ -383,10 +395,6 @@ describe('before loading', () => {
       assert(object)
       assert(object.context)
       assert(!object.context.amp)
-    })
-
-    describe('failed initializations', () => {
-      it.skip('should add failedInitializations as part of _metadata object if this.analytics.failedInitilizations is not empty', () => {})
     })
 
     describe('unbundling', () => {