+++ packages/browser/CHANGELOG.md
@@ -1,5 +1,82 @@
 # @segment/analytics-next
 
+## 1.53.2
+
+### Patch Changes
+
+- [#888](https://github.com/segmentio/analytics-next/pull/888) [`f3183f2`](https://github.com/segmentio/analytics-next/commit/f3183f2163d93f98ce2f8dd7830d2bca3b47f537) Thanks [@danieljackins](https://github.com/danieljackins)! - Fix query string parsing bug that was causing events containing the 'search' property with a non string value to be dropped
+
+* [#891](https://github.com/segmentio/analytics-next/pull/891) [`e0c7792`](https://github.com/segmentio/analytics-next/commit/e0c7792924e6f7ba0a36c200c5468141de3b9320) Thanks [@zikaari](https://github.com/zikaari)! - Fix cookie write error
+
+## 1.53.1
+
+### Patch Changes
+
+- [#900](https://github.com/segmentio/analytics-next/pull/900) [`9c8b609`](https://github.com/segmentio/analytics-next/commit/9c8b609f500d2338fdbf2572639241ffdfd27f8f) Thanks [@silesky](https://github.com/silesky)! - Add consent info to typescript types
+
+* [#896](https://github.com/segmentio/analytics-next/pull/896) [`48ce3ec`](https://github.com/segmentio/analytics-next/commit/48ce3ecbdea9361dde90481d7c50207613602d52) Thanks [@oscb](https://github.com/oscb)! - Added support for `screen` events for Segment destination
+
+## 1.53.0
+
+### Minor Changes
+
+- [#870](https://github.com/segmentio/analytics-next/pull/870) [`f23f3be`](https://github.com/segmentio/analytics-next/commit/f23f3bec9a774d9e6df8e243ab8e2711bacedd37) Thanks [@silesky](https://github.com/silesky)! - Add updateCDNSettings option
+
+* [#864](https://github.com/segmentio/analytics-next/pull/864) [`6cba535`](https://github.com/segmentio/analytics-next/commit/6cba5356c5f751e3edb30f7f524d4498600895b0) Thanks [@danieljackins](https://github.com/danieljackins)! - Add Client Hints API support
+
+### Patch Changes
+
+- Updated dependencies [[`6cba535`](https://github.com/segmentio/analytics-next/commit/6cba5356c5f751e3edb30f7f524d4498600895b0)]:
+  - @segment/analytics-core@1.3.0
+
+## 1.52.0
+
+### Minor Changes
+
+- [#867](https://github.com/segmentio/analytics-next/pull/867) [`2b3e5e4`](https://github.com/segmentio/analytics-next/commit/2b3e5e470b39ba6e23b90a8b2e61c6ed22b0fd6c) Thanks [@zikaari](https://github.com/zikaari)! - Device mode destination filters will now filter properties within arrays, just like they do in cloud mode
+
+### Patch Changes
+
+- [#861](https://github.com/segmentio/analytics-next/pull/861) [`99402e9`](https://github.com/segmentio/analytics-next/commit/99402e93902b6e5d02b8abe6944cc2f87255ca41) Thanks [@chrisradek](https://github.com/chrisradek)! - Fixes issue related to how retried events are stored in localStorage to prevent analytics.js from reading events for a different writeKey when that writeKey is used on the same domain as the current analytics.js.
+
+## 1.51.7
+
+### Patch Changes
+
+- Updated dependencies [[`7dcafa2`](https://github.com/segmentio/analytics-next/commit/7dcafa29cbce86d8c3d3c829c3ba7c22148949fc)]:
+  - @segment/analytics-core@1.2.5
+
+## 1.51.6
+
+### Patch Changes
+
+- [#842](https://github.com/segmentio/analytics-next/pull/842) [`2b71c10`](https://github.com/segmentio/analytics-next/commit/2b71c102519c12b056b596950f17e5c95fe5e41b) Thanks [@silesky](https://github.com/silesky)! - Fix 'Promise is undefined' issue with ie11 polyfill
+
+## 1.51.5
+
+### Patch Changes
+
+- [#838](https://github.com/segmentio/analytics-next/pull/838) [`55a48a0`](https://github.com/segmentio/analytics-next/commit/55a48a0d1e3589fc6b4896e94b79c857cabf1006) Thanks [@silesky](https://github.com/silesky)! - Refactor page enrichment to only call page defaults once, and simplify logic
+
+* [#839](https://github.com/segmentio/analytics-next/pull/839) [`fdc004b`](https://github.com/segmentio/analytics-next/commit/fdc004bebaa48206c710aab703b164b41e8fa984) Thanks [@silesky](https://github.com/silesky)! - Fixes a utm-parameter parsing bug where overridden page.search properties would not be reflected in the context.campaign object
+
+  ```ts
+  analytics.page(undefined, undefined, {search: "?utm_source=123&utm_content=content" )
+  analytics.track("foo", {url: "....", search: "?utm_source=123&utm_content=content" )
+
+  // should result in a context.campaign of:
+  { source: 123, content: 'content'}
+  ```
+
+## 1.51.4
+
+### Patch Changes
+
+- [#835](https://github.com/segmentio/analytics-next/pull/835) [`9353e09`](https://github.com/segmentio/analytics-next/commit/9353e0999f942db33066e337e9742ac2e96716b7) Thanks [@silesky](https://github.com/silesky)! - Refactor shared validation logic. Create granular error message if user ID does not match string type.
+
+- Updated dependencies [[`9353e09`](https://github.com/segmentio/analytics-next/commit/9353e0999f942db33066e337e9742ac2e96716b7)]:
+  - @segment/analytics-core@1.2.4
+
 ## 1.51.3
 
 ### Patch Changes